skipping shippable jdk

+++ shipctl/x86_64/Ubuntu_14.04/install.sh
@@ -17,7 +17,12 @@ echo "Installing shippable_replace"
 cp $SRC_DIR/shippable_replace /usr/local/bin/shippable_replace
 
 echo "Installing shippable_jdk"
-cp $SRC_DIR/shippable_jdk /usr/local/bin/shippable_jdk
+readonly shippable_jdk_location="/usr/local/bin/shippable_jdk"
+if [ if -f "$shippable_jdk_location" ]; then
+  echo "shippable_jdk already installed on the image, skipping"
+else
+  cp $SRC_DIR/shippable_jdk /usr/local/bin/shippable_jdk
+fi
 
 echo "Installing shipctl"
 cp $SRC_DIR/shipctl /usr/local/bin/shipctl
